rename and move 3 Services to ApiServices where they belong
