--- src/main/scala/org/broadinstitute/dsde/firecloud/service/HealthService.scala
@@ -1,19 +1,0 @@
-package org.broadinstitute.dsde.firecloud.service
-
-import org.slf4j.LoggerFactory
-import akka.http.scaladsl.model.StatusCodes._
-import akka.http.scaladsl.server.Route
-
-import scala.concurrent.ExecutionContext
-
-trait HealthService extends FireCloudDirectives {
-
-  implicit val executionContext: ExecutionContext
-  lazy val log = LoggerFactory.getLogger(getClass)
-
-  val healthServiceRoutes: Route = {
-    path("health") { complete(OK) } ~
-    path("error") { complete (ServiceUnavailable) }
-  }
-
-}
